Initial Version
This should resolve your issue, the reason your code was not working is because you kept getting that annoying signature setup prompt, this can be resolved by setting the Display Alerts property of the Excel object to false like so ;$Excel.DisplayAlerts = $False

This will associate some default behaviour to our excel instance so be careful to either re-enable this if you require any prompts etc. such as those that are produced by the Quit() method etc.