Custom signature location
"any ideas on setting position. I have it working right now but have to create it as a macro and then call it. It works but I assume PS surely can do it with less code."

Solution, choose a cell location for the signature to be placed, call the .Select() method on our cell range then insert the singature.